+++ src/lib.rs
@@ -8,9 +8,10 @@ extern crate static_assertions;
 
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
-use hyper::client::connect::HttpConnector;
 #[cfg(feature = "tls")]
 use hyper_tls::HttpsConnector;
+use hyper_util::client::legacy::{connect::HttpConnector, Client as HyperClient};
+use hyper_util::rt::TokioExecutor;
 
 pub use clickhouse_derive::Row;
 
@@ -33,6 +34,7 @@ mod buflist;
 mod compression;
 mod cursor;
 mod http_client;
+mod request_body;
 mod response;
 mod row;
 mod rowbinary;
@@ -46,10 +48,9 @@ const TCP_KEEPALIVE: Duration = Duration::from_secs(60);
 const POOL_IDLE_TIMEOUT: Duration = Duration::from_secs(2);
 
 /// A client containing HTTP pool.
-/// Can be created by using `Client::default()` or [`Client::with_http_client`].
 #[derive(Clone)]
 pub struct Client {
-    client: Arc<dyn HttpClient>,
+    http: Arc<dyn HttpClient>,
 
     url: String,
     database: Option<String>,
@@ -73,7 +74,7 @@ impl Default for Client {
             connector
         });
 
-        let client = hyper::Client::builder()
+        let client = HyperClient::builder(TokioExecutor::new())
             .pool_idle_timeout(POOL_IDLE_TIMEOUT)
             .build(connector);
 
@@ -83,10 +84,11 @@ impl Default for Client {
 
 impl Client {
     /// Creates a new client with a specified underlying HTTP client.
-    /// Now only [`hyper::Client`] is supported.
+    ///
+    /// See [`HttpClient`] for details.
     pub fn with_http_client(client: impl HttpClient) -> Self {
         Self {
-            client: Arc::new(client),
+            http: Arc::new(client),
             url: String::new(),
             database: None,
             user: None,